Create order table
A query to return the 10 earliest  orders in the orders table. Include the id, occurred_at and total_amt_usd.
A query to return the top 5 orders in the terms of largest total_amt_usd. Include the id, account id and total_amt_usd.
A query to return the lowest 20 orders in terms of smallest total_amt_usd. Include the id , account_id and total_amt_usd