Create Join
this result is the same as if you switched the tables in the FROM and JOIN. Additionally, which side of the = a column is listed doesn't matter.
Join is used with On. Join is used to pull any column or table to any other column or table if it has any relationship. To show a same column in multiple table we can use On